trie-db: rename core_ to rstd
