+++ trie-db/src/recorder.rs
@@ -72,7 +72,7 @@ impl<HO: Copy> Recorder<HO> {
 
 	/// Drain all visited records.
 	pub fn drain(&mut self) -> Vec<Record<HO>> {
-		crate::core_::mem::replace(&mut self.nodes, Vec::new())
+		crate::rstd::mem::replace(&mut self.nodes, Vec::new())
 	}
 }
